Fixes the maximum transmissions per batch, increases the transactions per block

+++ ledger/narwhal/subdag/src/bytes.rs
@@ -27,7 +27,7 @@ impl<N: Network> FromBytes for Subdag<N> {
         // Read the number of rounds.
         let num_rounds = u32::read_le(&mut reader)?;
         // Ensure the number of rounds is within bounds.
-        if num_rounds as usize > Self::MAX_ROUNDS {
+        if num_rounds as u64 > Self::MAX_ROUNDS {
             return Err(error(format!("Number of rounds ({num_rounds}) exceeds the maximum ({})", Self::MAX_ROUNDS)));
         }
         // Read the round certificates.
@@ -36,9 +36,9 @@ impl<N: Network> FromBytes for Subdag<N> {
             // Read the round.
             let round = u64::read_le(&mut reader)?;
             // Read the number of certificates.
-            let num_certificates = u32::read_le(&mut reader)?;
+            let num_certificates = u16::read_le(&mut reader)?;
             // Ensure the number of certificates is within bounds.
-            if num_certificates as usize > BatchHeader::<N>::MAX_CERTIFICATES {
+            if num_certificates > BatchHeader::<N>::MAX_CERTIFICATES {
                 return Err(error(format!(
                     "Number of certificates ({num_certificates}) exceeds the maximum ({})",
                     BatchHeader::<N>::MAX_CERTIFICATES
@@ -70,7 +70,7 @@ impl<N: Network> ToBytes for Subdag<N> {
             // Write the round.
             round.write_le(&mut writer)?;
             // Write the number of certificates.
-            u32::try_from(certificates.len()).map_err(error)?.write_le(&mut writer)?;
+            u16::try_from(certificates.len()).map_err(error)?.write_le(&mut writer)?;
             // Write the certificates.
             for certificate in certificates {
                 // Write the certificate.
